systems-engineering/duckdb-log-processing/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aider-chat==0.69.1
aiohappyeyeballs==2.4.4
    # via
    #   aider-chat
    #   aiohttp
aiohttp==3.11.10
    # via
    #   aider-chat
    #   litellm
aiosignal==1.3.1
    # via
    #   aider-chat
    #   aiohttp
annotated-types==0.7.0
    # via
    #   aider-chat
    #   pydantic
anyio==4.7.0
    # via
    #   aider-chat
    #   httpx
    #   openai
    #   watchfiles
attrs==24.2.0
    # via
    #   aider-chat
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via
    #   aider-chat
    #   posthog
beautifulsoup4==4.12.3
    # via aider-chat
certifi==2024.8.30
    # via
    #   aider-chat
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   aider-chat
    #   sounddevice
    #   soundfile
charset-normalizer==3.4.0
    # via
    #   aider-chat
    #   requests
click==8.1.7
    # via
    #   aider-chat
    #   litellm
configargparse==1.7
    # via aider-chat
diff-match-patch==20241021
    # via aider-chat
diskcache==5.6.3
    # via aider-chat
distro==1.9.0
    # via
    #   aider-chat
    #   openai
filelock==3.16.1
    # via
    #   aider-chat
    #   huggingface-hub
flake8==7.1.1
    # via aider-chat
frozenlist==1.5.0
    # via
    #   aider-chat
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via
    #   aider-chat
    #   huggingface-hub
gitdb==4.0.11
    # via
    #   aider-chat
    #   gitpython
gitpython==3.1.43
    # via aider-chat
grep-ast==0.4.1
    # via aider-chat
h11==0.14.0
    # via
    #   aider-chat
    #   httpcore
httpcore==1.0.7
    # via
    #   aider-chat
    #   httpx
httpx==0.27.2
    # via
    #   aider-chat
    #   litellm
    #   openai
huggingface-hub==0.26.5
    # via
    #   aider-chat
    #   tokenizers
idna==3.10
    # via
    #   aider-chat
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.2.1
    # via
    #   aider-chat
    #   litellm
importlib-resources==6.4.5
    # via aider-chat
jinja2==3.1.4
    # via
    #   aider-chat
    #   litellm
jiter==0.8.0
    # via
    #   aider-chat
    #   openai
json5==0.10.0
    # via aider-chat
jsonschema==4.23.0
    # via
    #   aider-chat
    #   litellm
jsonschema-specifications==2024.10.1
    # via
    #   aider-chat
    #   jsonschema
litellm==1.53.9
    # via aider-chat
markdown-it-py==3.0.0
    # via
    #   aider-chat
    #   rich
markupsafe==3.0.2
    # via
    #   aider-chat
    #   jinja2
mccabe==0.7.0
    # via
    #   aider-chat
    #   flake8
mdurl==0.1.2
    # via
    #   aider-chat
    #   markdown-it-py
mixpanel==4.10.1
    # via aider-chat
monotonic==1.6
    # via
    #   aider-chat
    #   posthog
multidict==6.1.0
    # via
    #   aider-chat
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via aider-chat
numpy==1.26.4
    # via
    #   aider-chat
    #   scipy
openai==1.57.0
    # via
    #   aider-chat
    #   litellm
packaging==24.2
    # via
    #   aider-chat
    #   huggingface-hub
pathspec==0.12.1
    # via
    #   aider-chat
    #   grep-ast
pexpect==4.9.0
    # via aider-chat
pillow==10.4.0
    # via aider-chat
pip==24.3.1
    # via aider-chat
posthog==3.7.4
    # via aider-chat
prompt-toolkit==3.0.48
    # via aider-chat
propcache==0.2.1
    # via
    #   aider-chat
    #   aiohttp
    #   yarl
psutil==6.1.0
    # via aider-chat
ptyprocess==0.7.0
    # via
    #   aider-chat
    #   pexpect
pycodestyle==2.12.1
    # via
    #   aider-chat
    #   flake8
pycparser==2.22
    # via
    #   aider-chat
    #   cffi
pydantic==2.10.3
    # via
    #   aider-chat
    #   litellm
    #   openai
pydantic-core==2.27.1
    # via
    #   aider-chat
    #   pydantic
pydub==0.25.1
    # via aider-chat
pyflakes==3.2.0
    # via
    #   aider-chat
    #   flake8
pygments==2.18.0
    # via
    #   aider-chat
    #   rich
pypandoc==1.14
    # via aider-chat
pyperclip==1.9.0
    # via aider-chat
python-dateutil==2.9.0.post0
    # via
    #   aider-chat
    #   posthog
python-dotenv==1.0.1
    # via
    #   aider-chat
    #   litellm
pyyaml==6.0.2
    # via
    #   aider-chat
    #   huggingface-hub
referencing==0.35.1
    # via
    #   aider-chat
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   aider-chat
    #   tiktoken
requests==2.32.3
    # via
    #   aider-chat
    #   huggingface-hub
    #   litellm
    #   mixpanel
    #   posthog
    #   tiktoken
rich==13.9.4
    # via aider-chat
rpds-py==0.22.3
    # via
    #   aider-chat
    #   jsonschema
    #   referencing
scipy==1.13.1
    # via aider-chat
six==1.17.0
    # via
    #   aider-chat
    #   mixpanel
    #   posthog
    #   python-dateutil
smmap==5.0.1
    # via
    #   aider-chat
    #   gitdb
sniffio==1.3.1
    # via
    #   aider-chat
    #   anyio
    #   httpx
    #   openai
sounddevice==0.5.1
    # via aider-chat
soundfile==0.12.1
    # via aider-chat
soupsieve==2.6
    # via
    #   aider-chat
    #   beautifulsoup4
tiktoken==0.8.0
    # via
    #   aider-chat
    #   litellm
tokenizers==0.19.1
    # via
    #   aider-chat
    #   litellm
tqdm==4.67.1
    # via
    #   aider-chat
    #   huggingface-hub
    #   openai
tree-sitter==0.21.3
    # via
    #   aider-chat
    #   tree-sitter-languages
tree-sitter-languages==1.10.2
    # via
    #   aider-chat
    #   grep-ast
typing-extensions==4.12.2
    # via
    #   aider-chat
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via
    #   aider-chat
    #   mixpanel
    #   requests
watchfiles==1.0.0
    # via aider-chat
wcwidth==0.2.13
    # via
    #   aider-chat
    #   prompt-toolkit
yarl==1.18.3
    # via
    #   aider-chat
    #   aiohttp
zipp==3.21.0
    # via
    #   aider-chat
    #   importlib-metadata
